refactor ♻️: Update QR code generation module
- Replaced outdated QR code generation module with updated version.
- Renames working file 'qrnew.py' back to 'qr_generator.py'
- Reverts import to reflect qr_generator and not qrnew
