+++ src/core/qr_generator.py
@@ -10,9 +10,11 @@ Copyright (c) 2024, LyAhn
 This code is licensed under the GPL-3.0 license (see LICENSE.txt for details)
 """
 import os
+import tempfile
 import io
 import json
 import cv2
+from MyQR import myqr
 import numpy as np
 from PIL import Image
 import qrcode
@@ -23,13 +25,14 @@ from PySide6.QtCore import Qt
 from PySide6.QtGui import QImage, QPixmap, QKeySequence, QShortcut, QIcon
 from PySide6.QtWidgets import QGraphicsScene, QPushButton, QSizePolicy, QScrollArea, QWidget
 from src.utils.templateEditor import JSONEditorDialog
-from PySide6.QtCore import QThreadPool, QRunnable, Slot, Signal, QObject
+from PySide6.QtCore import QThreadPool, QRunnable, Slot, Signal, QObject, QTimer
+
 
 class WorkerSignals(QObject):
     finished = Signal(object)
     error = Signal(tuple)
 
-class QRGeneratorWorker(QRunnable)
+class QRGeneratorWorker(QRunnable):
     def __init__(self, fn, *args, **kwargs):
         super().__init__()
         self.fn = fn
@@ -57,6 +60,9 @@ class QRGenerator:
         self.create_shortcut()
         self.editor = JSONEditorDialog
         self.threadpool = QThreadPool()
+        self.debounce_timer = QTimer()
+        self.debounce_timer.setSingleShot(True)
+        self.debounce_timer.timeout.connect(self.generate_qr_code_debounced)
 
     def setup_connections(self):
         self.ui.qrGenButton.clicked.connect(self.preview_qr_code)
@@ -72,6 +78,7 @@ class QRGenerator:
         self.ui.qrOutputView.mousePressEvent = self.on_preview_clicked
 
         #self.ui.qrTextInput.textChanged.connect(self.preview_qr_code) # remove this comment if you want live QR code preview while typing - Not advised
+        self.ui.qrTextInput.textChanged.connect(self.debounce_preview_qr_code)
         self.ui.qrCodeSize.valueChanged.connect(self.preview_qr_code)
         self.ui.qrBorderSize.valueChanged.connect(self.preview_qr_code)
         self.ui.qrErrorCorrectList.currentIndexChanged.connect(self.preview_qr_code)
@@ -79,11 +86,28 @@ class QRGenerator:
         self.ui.qrBgColourInput.textChanged.connect(self.preview_qr_code)
         self.ui.qrCodeColourInput.textChanged.connect(self.preview_qr_code)
         self.ui.actionQRTemplateEditor.triggered.connect(self.open_json_editor)
+        self.ui.qrUseArtisticCheck.stateChanged.connect(self.preview_qr_code)
+        self.ui.qrColorizedCheck.stateChanged.connect(self.preview_qr_code)
+        #self.ui.qrBgImageInput.textChanged.connect(self.preview_qr_code)
+
+    # Testing
+
+    def debounce_preview_qr_code(self):
+        self.debounce_timer.start(150)  # 300 ms debounce time default
+
+    def generate_qr_code_debounced(self):
+        worker = QRGeneratorWorker(self.generate_qr_code)
+        worker.signals.finished.connect(self.display_qr_code)
+        worker.signals.error.connect(self.handle_error)
+        self.threadpool.start(worker)
+
+    # End Testing
+
 
     async def generate_qr_async(self):
         loop = asyncio.get_event_loop()
         return await loop.run_in_executor(None, self.generate_qr_code)
-
+        
     def open_json_editor(self):
         editor = JSONEditorDialog('resources/qr_templates.json', self.ui)
         if editor.exec() == QDialog.Accepted:
@@ -96,12 +120,15 @@ class QRGenerator:
     def on_preview_clicked(self, event):
         if event.button() == Qt.LeftButton:
             self.show_preview_window()
-    
+
+    # def preview_qr_code(self):
+    #     worker = QRGeneratorWorker(self.generate_qr_code)
+    #     worker.signals.finished.connect(self.display_qr_code)
+    #     worker.signals.error.connect(self.handle_error)
+    #     self.threadpool.start(worker)
+
     def preview_qr_code(self):
-        worker = QRGeneratorWorker(self.generate_qr_code)
-        worker.signals.finished.connect(self.display_qr_code)
-        worker.signals.error.connect(self.handle_error)
-        self.threadpool.start(worker)
+        self.generate_qr_code_debounced()
 
     def handle_error(self, error):
         QMessageBox.critical(None, "Error", f"An error occured: {error[0]}")
@@ -139,11 +166,142 @@ class QRGenerator:
     #     qr.add_data(qr_data)
     #     qr.make(fit=True)
 
+    # def generate_qr_code(self):
+    #     qr_data = self.ui.qrTextInput.toPlainText().strip()
+    #     if not qr_data:
+    #         return None
+        
+    #     use_artistic = self.ui.qrUseArtisticCheck.isChecked()
+        
+    #     if use_artistic:
+    #         return self.generate_artistic_qr(qr_data)
+    #     else:
+    #         return self.generate_standard_qr(qr_data)
+
     def generate_qr_code(self):
         qr_data = self.ui.qrTextInput.toPlainText().strip()
-        if not qr_data:
+        if not qr_data or len(qr_data) < 2:
             return None
 
+        use_artistic = self.ui.qrUseArtisticCheck.isChecked()
+
+        if use_artistic:
+            return self.generate_artistic_qr(qr_data)
+        else:
+            # Check if the data is purely numeric
+            if qr_data.isdigit():
+                mode = 'numeric'
+            else:
+                mode = 'binary'  # Use byte mode for non-numeric data
+
+            qr = qrcode.QRCode(
+                version=self.ui.qrCodeSize.value(),
+                error_correction=self.get_error_correction(),
+                box_size=10,
+                border=self.ui.qrBorderSize.value(),
+            )
+            qr.add_data(qr_data, optimize=0)  # Disable automatic mode optimization
+            qr.make(fit=True)
+
+            bg_color = self.get_color_tuple(self.ui.qrBgColourInput.text(), (255, 255, 255))
+            fg_color = self.get_color_tuple(self.ui.qrCodeColourInput.text(), (0, 0, 0))
+            qr_image = qr.make_image(fill_color=fg_color, back_color=bg_color)
+            qr_image = qr_image.convert('RGB')
+
+            qr_array = np.array(qr_image)
+            qr_cv = cv2.cvtColor(qr_array, cv2.COLOR_RGB2BGR)
+            target_size = (1024, 1024)
+            qr_code_resized = cv2.resize(qr_cv, target_size, interpolation=cv2.INTER_AREA)
+
+            logo_path = self.ui.qrLogoInput.text()
+            if logo_path and os.path.isfile(logo_path):
+                qr_code_resized = self.add_logo_to_qr(qr_code_resized, logo_path)
+
+            return Image.fromarray(cv2.cvtColor(qr_code_resized, cv2.COLOR_BGR2RGB))
+
+    # def generate_artistic_qr(self, qr_data):
+    #     version = self.ui.qrCodeSize.value()
+    #     error_correction = self.get_error_correction_level()
+    #     picture = self.ui.qrLogoInput.text()
+    #     #picture = self.ui.qrBgImageInput.text()
+    #     colorized = self.ui.qrColorizedCheck.isChecked()
+    #     border_size = self.ui.qrBorderSize.value()
+        
+    #     save_name = "temp_artistic_qr.png"
+    #     version, level, qr_name = myqr.run(
+    #         qr_data,
+    #         version=version,
+    #         level=error_correction,
+    #         picture=picture,
+    #         colorized=colorized,
+    #         save_name=save_name
+    #     )
+        
+    #     # Open the generated QR code
+    #     qr_image = Image.open(save_name)
+        
+    #     # Crop the white border
+    #     bbox = qr_image.getbbox()
+    #     cropped_qr = qr_image.crop(bbox)
+        
+    #     # Create a new image with desired border
+    #     qr_size = cropped_qr.size[0]
+    #     new_size = qr_size + 2 * border_size
+    #     bg_color = self.get_color_tuple(self.ui.qrBgColourInput.text(), (255, 255, 255))
+    #     new_image = Image.new('RGB', (new_size, new_size), bg_color)
+        
+    #     # Paste the cropped QR code onto the new image
+    #     new_image.paste(cropped_qr, (border_size, border_size))
+        
+    #     # Remove the temporary file
+    #     os.remove(save_name)
+        
+    #     return new_image
+    def generate_artistic_qr(self, qr_data):
+        try:
+            version = self.ui.qrCodeSize.value()
+            error_correction = self.get_error_correction_level()
+            picture = self.ui.qrLogoInput.text()
+            colorized = self.ui.qrColorizedCheck.isChecked()
+            border_size = self.ui.qrBorderSize.value()
+
+            # Use a temporary directory to save the file
+            with tempfile.TemporaryDirectory() as temp_dir:
+                save_name = os.path.join(temp_dir, "temp_artistic_qr.png")
+                
+                version, level, qr_name = myqr.run(
+                    qr_data,
+                    version=version,
+                    level=error_correction,
+                    picture=picture,
+                    colorized=colorized,
+                    save_name=save_name
+                )
+                
+                # Open the generated QR code
+                qr_image = Image.open(save_name)
+                
+                # Crop the white border
+                bbox = qr_image.getbbox()
+                cropped_qr = qr_image.crop(bbox)
+                
+                # Create a new image with desired border
+                qr_size = cropped_qr.size[0]
+                new_size = qr_size + 2 * border_size
+                bg_color = self.get_color_tuple(self.ui.qrBgColourInput.text(), (255, 255, 255))
+                new_image = Image.new('RGB', (new_size, new_size), bg_color)
+                
+                # Paste the cropped QR code onto the new image
+                new_image.paste(cropped_qr, (border_size, border_size))
+                
+                return new_image
+
+        except Exception as e:
+            print(f"Error in generate_artistic_qr: {str(e)}")
+            # Return a default QR code or None
+            return self.generate_standard_qr(qr_data)
+
+    def generate_standard_qr(self, qr_data):
         qr = qrcode.QRCode(
             version=self.ui.qrCodeSize.value(),
             error_correction=self.get_error_correction(),
@@ -155,64 +313,51 @@ class QRGenerator:
 
         bg_color = self.get_color_tuple(self.ui.qrBgColourInput.text(), (255, 255, 255))
         fg_color = self.get_color_tuple(self.ui.qrCodeColourInput.text(), (0, 0, 0))
+
         qr_image = qr.make_image(fill_color=fg_color, back_color=bg_color)
         qr_image = qr_image.convert('RGB')
 
         qr_array = np.array(qr_image)
         qr_cv = cv2.cvtColor(qr_array, cv2.COLOR_RGB2BGR)
-
         target_size = (1024, 1024)
         qr_code_resized = cv2.resize(qr_cv, target_size, interpolation=cv2.INTER_AREA)
 
         logo_path = self.ui.qrLogoInput.text()
         if logo_path and os.path.isfile(logo_path):
-            logo = Image.open(logo_path).convert('RGBA')
-            if logo is not None:
-                logo_size = min(qr_code_resized.shape[0], qr_code_resized.shape[1]) // 3
-
-                if self.ui.qrAspectRatioCheck.isChecked():
-                    # Preserve aspect ratio
-                    logo.thumbnail((logo_size, logo_size), Image.LANCZOS)
-                else:
-                    # Resize without preserving aspect ratio
-                    logo = logo.resize((logo_size, logo_size), Image.LANCZOS)
-
-                if self.ui.qrAddBGCheck.isChecked():
-                    bg_size = (logo_size, logo_size)
-                    logo_bg = Image.new('RGBA', bg_size, (255, 255, 255, 255))
-                    logo_pos = ((bg_size[0] - logo.size[0]) // 2,
-                                (bg_size[1] - logo.size[1]) // 2)
-                    logo_bg.paste(logo, logo_pos, logo)
-                    logo = logo_bg
-
-                logo_np = np.array(logo)
-                logo_cv = cv2.cvtColor(logo_np, cv2.COLOR_RGBA2BGRA)
-
-                top_left_x = (qr_code_resized.shape[1] - logo_cv.shape[1]) // 2
-                top_left_y = (qr_code_resized.shape[0] - logo_cv.shape[0]) // 2
-
-                if logo_cv.shape[2] == 4:
-                    alpha = logo_cv[:, :, 3] / 255.0
-                    alpha = np.expand_dims(alpha, axis=2)
-                    rgb = cv2.cvtColor(logo_cv, cv2.COLOR_BGRA2BGR)
-
-                    roi = qr_code_resized[top_left_y:top_left_y+logo_cv.shape[0], 
-                                        top_left_x:top_left_x+logo_cv.shape[1]]
-                    qr_code_resized[top_left_y:top_left_y+logo_cv.shape[0], 
-                                    top_left_x:top_left_x+logo_cv.shape[1]] = \
-                        (1 - alpha) * roi + alpha * rgb
+            qr_code_resized = self.add_logo_to_qr(qr_code_resized, logo_path)
 
         return Image.fromarray(cv2.cvtColor(qr_code_resized, cv2.COLOR_BGR2RGB))
 
-    def generate_and_display_qr(self):
-        qr_data = self.ui.qrTextInput.toPlainText().strip()
-        if qr_data:
-            qr_image = self.generate_qr_code()
-            if qr_image:
-                self.display_qr_code(qr_image)
-        else:
-            self.clear_qr_display()
-            self.current_qr_image = None
+    def add_logo_to_qr(self, qr_code_resized, logo_path):
+        logo = Image.open(logo_path).convert('RGBA')
+        if logo is not None:
+            logo_size = min(qr_code_resized.shape[0], qr_code_resized.shape[1]) // 3
+            if self.ui.qrAspectRatioCheck.isChecked():
+                logo.thumbnail((logo_size, logo_size), Image.LANCZOS)
+            else:
+                logo = logo.resize((logo_size, logo_size), Image.LANCZOS)
+
+            if self.ui.qrAddBGCheck.isChecked():
+                bg_size = (logo_size, logo_size)
+                logo_bg = Image.new('RGBA', bg_size, (255, 255, 255, 255))
+                logo_pos = ((bg_size[0] - logo.size[0]) // 2, (bg_size[1] - logo.size[1]) // 2)
+                logo_bg.paste(logo, logo_pos, logo)
+                logo = logo_bg
+
+            logo_np = np.array(logo)
+            logo_cv = cv2.cvtColor(logo_np, cv2.COLOR_RGBA2BGRA)
+            top_left_x = (qr_code_resized.shape[1] - logo_cv.shape[1]) // 2
+            top_left_y = (qr_code_resized.shape[0] - logo_cv.shape[0]) // 2
+
+            if logo_cv.shape[2] == 4:
+                alpha = logo_cv[:, :, 3] / 255.0
+                alpha = np.expand_dims(alpha, axis=2)
+                rgb = cv2.cvtColor(logo_cv, cv2.COLOR_BGRA2BGR)
+                roi = qr_code_resized[top_left_y:top_left_y+logo_cv.shape[0], top_left_x:top_left_x+logo_cv.shape[1]]
+                qr_code_resized[top_left_y:top_left_y+logo_cv.shape[0], top_left_x:top_left_x+logo_cv.shape[1]] = \
+                    (1 - alpha) * roi + alpha * rgb
+
+        return qr_code_resized
 
     def get_color_tuple(self, color_string, default):
         try:
@@ -310,16 +455,11 @@ class QRGenerator:
         try:
             with open('resources/qr_templates.json', 'r') as file:
                 data = json.load(file)
-                self.qr_templates = data['qr_code_types']
-
-            self.ui.qrTemplates.clear()
-            self.ui.qrTemplates.addItem("Select a template")
-            for template in self.qr_templates:
-                self.ui.qrTemplates.addItem(template['name'])
+            self.qr_templates = data['qr_code_types']
+            self.populate_qr_templates()
         except FileNotFoundError:
             print("Error: 'resources/qr_templates.json' file not found.")
             QMessageBox.warning(None, "Warning", "Error: 'resources/qr_templates.json' file not found.\nRedownload the template file and try again.")
-
             self.qr_templates = []
         except json.JSONDecodeError:
             print("Error: Invalid JSON format in 'resources/qr_templates.json'.")
@@ -415,18 +555,87 @@ class QRGenerator:
 
             # Resize the pixmap to fit the label while preserving aspect ratio
             label.resizeEvent = lambda event: label.setPixmap(pixmap.scaled(
-                label.width(), label.height(),
-                Qt.KeepAspectRatio, Qt.SmoothTransformation
+                label.width(), label.height(), Qt.KeepAspectRatio, Qt.SmoothTransformation
             ))
 
             layout.addWidget(label)
-
             close_button = QPushButton("Close")
             close_button.clicked.connect(preview_dialog.close)
             layout.addWidget(close_button)
-
             preview_dialog.setLayout(layout)
             preview_dialog.exec()
         else:
-            # Only show a warning if there's no QR code and the user explicitly tries to preview
             QMessageBox.warning(None, "Warning", "No QR code has been generated yet.")
+
+    # def generate_artistic_qr(self, qr_data):
+        
+    #     version = self.ui.qrCodeSize.value()
+    #     error_correction = self.get_error_correction_level()
+    #     picture = self.ui.qrLogoInput.text()
+    #     #picture = self.ui.qrBgImageInput.text()
+    #     colorized = self.ui.qrColorizedCheck.isChecked()
+        
+    #     save_name = "temp_artistic_qr.png"
+    #     version, level, qr_name = myqr.run(
+    #         qr_data,
+    #         version=version,
+    #         level=error_correction,
+    #         picture=picture,
+    #         colorized=colorized,
+    #         save_name=save_name
+    #     )
+        
+    #     return Image.open(save_name)
+
+    def generate_artistic_qr(self, qr_data):
+        version = self.ui.qrCodeSize.value()
+        error_correction = self.get_error_correction_level()
+        picture = self.ui.qrLogoInput.text()
+        colorized = self.ui.qrColorizedCheck.isChecked()
+        border_size = self.ui.qrBorderSize.value()
+
+        with tempfile.TemporaryDirectory() as temp_dir:
+            save_name = os.path.join(temp_dir, "temp_artistic_qr.png")
+            try:
+                version, level, qr_name = myqr.run(
+                    qr_data,
+                    version=version,
+                    level=error_correction,
+                    picture=picture,
+                    colorized=colorized,
+                    save_name=save_name
+                )
+
+                # Open the generated QR code
+                qr_image = Image.open(save_name)
+
+                # Crop the white border
+                bbox = qr_image.getbbox()
+                cropped_qr = qr_image.crop(bbox)
+
+                # Create a new image with desired border
+                qr_size = cropped_qr.size[0]
+                new_size = qr_size + 2 * border_size
+                bg_color = self.get_color_tuple(self.ui.qrBgColourInput.text(), (255, 255, 255))
+                new_image = Image.new('RGB', (new_size, new_size), bg_color)
+
+                # Paste the cropped QR code onto the new image
+                new_image.paste(cropped_qr, (border_size, border_size))
+
+                return new_image
+
+            except Exception as e:
+                print(f"Error in generate_artistic_qr: {str(e)}")
+                # Return a default QR code or None
+                return self.generate_standard_qr(qr_data)
+
+    def get_error_correction_level(self):
+        error_correction = self.ui.qrErrorCorrectList.currentText()
+        if error_correction == "Low":
+            return 'L'
+        elif error_correction == "Medium":
+            return 'M'
+        elif error_correction == "Quartile":
+            return 'Q'
+        else:
+            return 'H'
